--- scale-decode/src/visitor/ext/mod.rs
@@ -1,20 +1,0 @@
-// Copyright (C) 2022 Parity Technologies (UK) Ltd. (admin@parity.io)
-// This file is a part of the scale-value crate.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//         http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//! Types associated with the `VisitorExt` trait methods.
-
-mod and_then;
-
-pub use and_then::{AndThen, AndThenFn};
